cleaned up the grouping_config.py file
